- NPCAddons + TimingManager and LagMeter | Minor Changes

+++ src/main/java/xyz/ufactions/crates/CratesModule.java
@@ -2,11 +2,14 @@ package xyz.ufactions.crates;
 
 import com.gmail.filoghost.holographicdisplays.api.Hologram;
 import com.gmail.filoghost.holographicdisplays.api.HologramsAPI;
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
-import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
+import xyz.ufactions.TabCompleteEvent;
 import xyz.ufactions.api.Module;
 import xyz.ufactions.crates.commands.CrateCommand;
 import xyz.ufactions.crates.files.CrateFiles;
@@ -15,6 +18,7 @@ import xyz.ufactions.crates.listeners.PlayerInteract;
 import xyz.ufactions.crates.managers.LocationManager;
 import xyz.ufactions.crates.objects.Crate;
 import xyz.ufactions.crates.utils.UtilChat;
+import xyz.ufactions.libs.F;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -24,8 +28,15 @@ public class CratesModule extends Module {
 
     private static HashSet<CrateHook> hooks = new HashSet<>();
 
+    private boolean holograms = false;
+
     public CratesModule(JavaPlugin plugin) {
         super("Crates", plugin);
+
+        holograms = Bukkit.getPluginManager().isPluginEnabled("HolographicDisplays");
+
+        if (holograms)
+            System.out.println("HolographicDisplays found! Crates hooking onto API");
 
         instance = this;
         prefix = "CustomCrates";
@@ -67,16 +78,18 @@ public class CratesModule extends Module {
     }
 
     private void loadHolograms() {
-        for (Hologram holo : HologramsAPI.getHolograms(Plugin)) {
-            holo.delete();
-        }
-        for (Crate crate : Crate.getCrates()) {
-            for (Location loc : new LocationManager().getLocations(crate)) {
-                loc.add(0.5, 1.7, 0.5);
-                Hologram holo = HologramsAPI.createHologram(Plugin, loc);
-                holo.insertTextLine(0, UtilChat.cc(crate.getDisplayName()));
-                holo.insertTextLine(1, UtilChat.cc("&7(&fLeft click &7to see rewards)"));
-                holo.insertTextLine(2, UtilChat.cc("&7(&fRight click &7with a key)"));
+        if (isHolograms()) {
+            for (Hologram holo : HologramsAPI.getHolograms(Plugin)) {
+                holo.delete();
+            }
+            for (Crate crate : Crate.getCrates()) {
+                for (Location loc : new LocationManager().getLocations(crate)) {
+                    loc.add(0.5, 1.7, 0.5);
+                    Hologram holo = HologramsAPI.createHologram(Plugin, loc);
+                    holo.insertTextLine(0, UtilChat.cc(crate.getDisplayName()));
+                    holo.insertTextLine(1, UtilChat.cc("&7(&fLeft click &7to see rewards)"));
+                    holo.insertTextLine(2, UtilChat.cc("&7(&fRight click &7with a key)"));
+                }
             }
         }
     }
@@ -99,5 +112,9 @@ public class CratesModule extends Module {
 
     public static HashSet<CrateHook> getHooks() {
         return hooks;
+    }
+
+    public boolean isHolograms() {
+        return holograms;
     }
 }
